Fix on real hardware!
Aside from a silly mistake, the main issue why this was crashing on
real hardware before was because the 3DS GPU apparently does not like
rendering using one shader program with a geo component and one without
on the same frame. Making a dummy geoshader for the normal sprite case
appears to fix it (not sure what performance cost this may entail).